requirements.txt
absl-py==1.4.0
aggdraw==1.3.16
aiohttp==3.8.4
aiosignal==1.3.1
altgraph==0.17.4
appdirs==1.4.4
astunparse==1.6.3
async-timeout==4.0.2
attrs==22.2.0
beautifulsoup4==4.12.2
blinker==1.6.2
blis==0.7.9
bs4==0.0.1
cachetools==5.3.1
catalogue==2.0.8
certifi==2022.12.7
chardet==4.0.0
charset-normalizer==3.1.0
chess==1.10.0
click==8.1.6
confection==0.1.0
contourpy==1.1.0
cycler==0.10.0
cymem==2.0.7
discord.py==2.2.2
EasyProcess==1.1
entrypoint2==1.1
fastai==2.7.12
fastcore==1.5.29
fastdownload==0.0.7
fastprogress==1.0.3
filelock==3.12.2
Flask==2.3.2
Flask-Cors==4.0.0
flatbuffers==23.5.26
fonttools==4.41.0
frozendict==2.3.10
frozenlist==1.3.3
gast==0.4.0
google-auth==2.22.0
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
grpcio==1.56.2
h5py==3.9.0
html5lib==1.1
idna==2.10
inflection==0.5.1
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.3.1
keras==2.13.1
keyboard==0.13.5
kiwisolver==1.4.4
langcodes==3.3.0
libclang==16.0.6
lxml==4.9.3
macholib==1.16.3
macos-notifications==0.1.6
Markdown==3.4.4
MarkupSafe==2.1.3
matplotlib==3.7.2
more-itertools==10.1.0
MouseInfo==0.1.3
mpmath==1.3.0
mss==9.0.1
multidict==6.0.4
multitasking==0.0.11
murmurhash==1.0.9
networkx==3.1
numpy==1.24.3
oauthlib==3.2.2
onnx==1.14.0
opencv-python==4.8.0.76
opencv-python-headless==4.8.0.74
opt-einsum==3.3.0
packaging==23.1
pandas==2.0.3
pathy==0.10.2
peewee==3.17.0
Pillow==9.5.0
playsound==1.3.0
plyer==2.1.0
preshed==3.0.8
protobuf==4.23.4
psutil==5.9.5
py-cpuinfo==9.0.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
PyAutoGUI==0.9.54
pydantic==1.10.11
pygame==2.5.2
PyGetWindow==0.0.9
pyinstaller==6.1.0
pyinstaller-hooks-contrib==2023.10
PyMsgBox==1.0.9
pyobjc==9.2
pyobjc-core==9.2
pyobjc-framework-Accessibility==9.2
pyobjc-framework-Accounts==9.2
pyobjc-framework-AddressBook==9.2
pyobjc-framework-AdServices==9.2
pyobjc-framework-AdSupport==9.2
pyobjc-framework-AppleScriptKit==9.2
pyobjc-framework-AppleScriptObjC==9.2
pyobjc-framework-ApplicationServices==9.2
pyobjc-framework-AppTrackingTransparency==9.2
pyobjc-framework-AudioVideoBridging==9.2
pyobjc-framework-AuthenticationServices==9.2
pyobjc-framework-AutomaticAssessmentConfiguration==9.2
pyobjc-framework-Automator==9.2
pyobjc-framework-AVFoundation==9.2
pyobjc-framework-AVKit==9.2
pyobjc-framework-AVRouting==9.2
pyobjc-framework-BackgroundAssets==9.2
pyobjc-framework-BusinessChat==9.2
pyobjc-framework-CalendarStore==9.2
pyobjc-framework-CallKit==9.2
pyobjc-framework-CFNetwork==9.2
pyobjc-framework-ClassKit==9.2
pyobjc-framework-CloudKit==9.2
pyobjc-framework-Cocoa==9.2
pyobjc-framework-Collaboration==9.2
pyobjc-framework-ColorSync==9.2
pyobjc-framework-Contacts==9.2
pyobjc-framework-ContactsUI==9.2
pyobjc-framework-CoreAudio==9.2
pyobjc-framework-CoreAudioKit==9.2
pyobjc-framework-CoreBluetooth==9.2
pyobjc-framework-CoreData==9.2
pyobjc-framework-CoreHaptics==9.2
pyobjc-framework-CoreLocation==9.2
pyobjc-framework-CoreMedia==9.2
pyobjc-framework-CoreMediaIO==9.2
pyobjc-framework-CoreMIDI==9.2
pyobjc-framework-CoreML==9.2
pyobjc-framework-CoreMotion==9.2
pyobjc-framework-CoreServices==9.2
pyobjc-framework-CoreSpotlight==9.2
pyobjc-framework-CoreText==9.2
pyobjc-framework-CoreWLAN==9.2
pyobjc-framework-CryptoTokenKit==9.2
pyobjc-framework-DataDetection==9.2
pyobjc-framework-DeviceCheck==9.2
pyobjc-framework-DictionaryServices==9.2
pyobjc-framework-DiscRecording==9.2
pyobjc-framework-DiscRecordingUI==9.2
pyobjc-framework-DiskArbitration==9.2
pyobjc-framework-DVDPlayback==9.2
pyobjc-framework-EventKit==9.2
pyobjc-framework-ExceptionHandling==9.2
pyobjc-framework-ExecutionPolicy==9.2
pyobjc-framework-ExtensionKit==9.2
pyobjc-framework-ExternalAccessory==9.2
pyobjc-framework-FileProvider==9.2
pyobjc-framework-FileProviderUI==9.2
pyobjc-framework-FinderSync==9.2
pyobjc-framework-FSEvents==9.2
pyobjc-framework-GameCenter==9.2
pyobjc-framework-GameController==9.2
pyobjc-framework-GameKit==9.2
pyobjc-framework-GameplayKit==9.2
pyobjc-framework-HealthKit==9.2
pyobjc-framework-ImageCaptureCore==9.2
pyobjc-framework-IMServicePlugIn==9.2
pyobjc-framework-InputMethodKit==9.2
pyobjc-framework-InstallerPlugins==9.2
pyobjc-framework-InstantMessage==9.2
pyobjc-framework-Intents==9.2
pyobjc-framework-IntentsUI==9.2
pyobjc-framework-IOBluetooth==9.2
pyobjc-framework-IOBluetoothUI==9.2
pyobjc-framework-IOSurface==9.2
pyobjc-framework-iTunesLibrary==9.2
pyobjc-framework-KernelManagement==9.2
pyobjc-framework-LatentSemanticMapping==9.2
pyobjc-framework-LaunchServices==9.2
pyobjc-framework-libdispatch==9.2
pyobjc-framework-libxpc==9.2
pyobjc-framework-LinkPresentation==9.2
pyobjc-framework-LocalAuthentication==9.2
pyobjc-framework-LocalAuthenticationEmbeddedUI==9.2
pyobjc-framework-MailKit==9.2
pyobjc-framework-MapKit==9.2
pyobjc-framework-MediaAccessibility==9.2
pyobjc-framework-MediaLibrary==9.2
pyobjc-framework-MediaPlayer==9.2
pyobjc-framework-MediaToolbox==9.2
pyobjc-framework-Metal==9.2
pyobjc-framework-MetalFX==9.2
pyobjc-framework-MetalKit==9.2
pyobjc-framework-MetalPerformanceShaders==9.2
pyobjc-framework-MetalPerformanceShadersGraph==9.2
pyobjc-framework-MetricKit==9.2
pyobjc-framework-MLCompute==9.2
pyobjc-framework-ModelIO==9.2
pyobjc-framework-MultipeerConnectivity==9.2
pyobjc-framework-NaturalLanguage==9.2
pyobjc-framework-NetFS==9.2
pyobjc-framework-Network==9.2
pyobjc-framework-NetworkExtension==9.2
pyobjc-framework-NotificationCenter==9.2
pyobjc-framework-OpenDirectory==9.2
pyobjc-framework-OSAKit==9.2
pyobjc-framework-OSLog==9.2
pyobjc-framework-PassKit==9.2
pyobjc-framework-PencilKit==9.2
pyobjc-framework-PHASE==9.2
pyobjc-framework-Photos==9.2
pyobjc-framework-PhotosUI==9.2
pyobjc-framework-PreferencePanes==9.2
pyobjc-framework-PushKit==9.2
pyobjc-framework-Quartz==9.2
pyobjc-framework-QuickLookThumbnailing==9.2
pyobjc-framework-ReplayKit==9.2
pyobjc-framework-SafariServices==9.2
pyobjc-framework-SafetyKit==9.2
pyobjc-framework-SceneKit==9.2
pyobjc-framework-ScreenCaptureKit==9.2
pyobjc-framework-ScreenSaver==9.2
pyobjc-framework-ScreenTime==9.2
pyobjc-framework-ScriptingBridge==9.2
pyobjc-framework-SearchKit==9.2
pyobjc-framework-Security==9.2
pyobjc-framework-SecurityFoundation==9.2
pyobjc-framework-SecurityInterface==9.2
pyobjc-framework-ServiceManagement==9.2
pyobjc-framework-SharedWithYou==9.2
pyobjc-framework-SharedWithYouCore==9.2
pyobjc-framework-ShazamKit==9.2
pyobjc-framework-Social==9.2
pyobjc-framework-SoundAnalysis==9.2
pyobjc-framework-Speech==9.2
pyobjc-framework-SpriteKit==9.2
pyobjc-framework-StoreKit==9.2
pyobjc-framework-SyncServices==9.2
pyobjc-framework-SystemConfiguration==9.2
pyobjc-framework-SystemExtensions==9.2
pyobjc-framework-ThreadNetwork==9.2
pyobjc-framework-UniformTypeIdentifiers==9.2
pyobjc-framework-UserNotifications==9.2
pyobjc-framework-UserNotificationsUI==9.2
pyobjc-framework-VideoSubscriberAccount==9.2
pyobjc-framework-VideoToolbox==9.2
pyobjc-framework-Virtualization==9.2
pyobjc-framework-Vision==9.2
pyobjc-framework-WebKit==9.2
pyparsing==2.4.7
pyperclip==1.8.2
PyRect==0.2.0
pyscreenshot==3.1
PyScreeze==0.1.29
pytesseract==0.3.10
python-dateutil==2.8.2
python-dotenv==1.0.0
pytweening==1.0.7
pytz==2023.3
PyYAML==6.0.1
Quandl==3.7.0
requests==2.31.0
requests-oauthlib==1.3.1
requests-toolbelt==1.0.0
roboflow==1.1.7
rsa==4.9
rubicon-objc==0.4.6
schedule==1.2.1
scikit-learn==1.3.0
scipy==1.11.1
seaborn==0.12.2
six==1.16.0
smart-open==6.3.0
soupsieve==2.5
spacy==3.6.0
spacy-legacy==3.0.12
spacy-loggers==1.0.4
srsly==2.4.7
supervision==0.14.0
sympy==1.12
tensorboard==2.13.0
tensorboard-data-server==0.7.1
tensorflow==2.13.0
tensorflow-estimator==2.13.0
tensorflow-io-gcs-filesystem==0.32.0
termcolor==2.3.0
thinc==8.1.10
threadpoolctl==3.2.0
torch==2.0.1
torch-tb-profiler==0.4.1
torchvision==0.15.2
tqdm==4.65.0
typer==0.9.0
typing_extensions==4.5.0
tzdata==2023.3
ultralytics==8.0.152
urllib3==1.26.16
visualkeras==0.0.2
wasabi==1.1.2
webencodings==0.5.1
Werkzeug==2.3.6
wrapt==1.15.0
yarl==1.8.2
yfinance==0.2.32
